Cleaned up + added service layer
